+++ src/main/java/qma/ajuda/AjudaRepository.java
@@ -1,8 +1,8 @@
 package qma.ajuda;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.transaction.Transactional;
 
@@ -10,26 +10,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @Transactional
-public class AjudaRepository {
+public interface AjudaRepository extends JpaRepository<Ajuda, Long>{
 	
-	private Map<Long, Ajuda> ajudas = new HashMap<Long, Ajuda>();
-	
-	
-	public Ajuda findById(Long id) {
-		if (ajudas.get(id) != null) {
-			return ajudas.get(id);
-		}
-		
-		throw new RuntimeException("Ajuda nao encontrada");
-	}
-	
-	public Ajuda save(Ajuda ajuda) {
-		ajudas.put(ajuda.getId(), ajuda);
-		return ajuda;
-	}
-	
-	public Collection<Ajuda> findAll(){
-		return ajudas.values();
-	}
+	Optional<Ajuda> findById(Long id);
 
 }
