+++ src/main/java/qma/aluno/AlunoRestController.java
@@ -1,6 +1,7 @@
 package qma.aluno;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping(value = "/")
+@CrossOrigin
 public class AlunoRestController {
 	
 	@Autowired
