+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-server.port = 9093
+server.port = 8090
